refactor: ThemePage 모듈화 및 코드 분리
- getProducts 함수 분리
- ProductSection 컴포넌트 분리
- ThemePage 컴포넌트를 모듈화하여 코드 가독성 향상
